Another big set of network changes
Many changes here. Game and server are now “bundled”. Big part of
ServerMessenger is a black box - convoluted code that I wrote to try
and verify user connections, but only half working. Basically, what we
have right now is the ability to start the server, then start up to 6
(right now game automatically starts at 2) players on different
machines, and have the first player begin his turn. Not much, but it
shouldn’t be too hard to go off of it.